MatchActivity refactored into fragment
